use int64 for size, fix glob install/update/erase, fix wrong exit on erase with multiple pkgs
(cherry picked from commit eb0c3ff38fc3e83a86ef091466608568b81b066a)

+++ common/prototypes.h
@@ -128,7 +128,7 @@ TDNFRightTrim(
 
 uint32_t
 TDNFUtilsFormatSize(
-    uint32_t unSize,
+    uint64_t unSize,
     char** ppszFormattedSize
     );
